pyproject.toml
[project]
name = "raumzeit-ki-agent"
version = "0.1.0"
description = "Telegram bot with Claude tool-use for Raumzeit API"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "litellm>=1.50.0",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
